Fix ultra_light_face_detector data feed issue

+++ modules/image/face_detection/ultra_light_fast_generic_face_detector_1mb_320/data_feed.py
@@ -34,7 +34,7 @@ def reader(images=None, paths=None):
             each = OrderedDict()
             assert os.path.isfile(
                 im_path), "The {} isn't a valid file path.".format(im_path)
-            im = cv2.imread(im_path).astype('float32')
+            im = cv2.imread(im_path)
             each['orig_im'] = im
             each['orig_im_shape'] = im.shape  # height, width, channel
             each['orig_im_path'] = im_path
